+++ packages/components/src/utils/property-checkers/index.ts
@@ -19,3 +19,5 @@ export const checkRequiredAndPattern = requiredAnd(checkPattern);
 export const checkRequiredAndType = requiredAnd(checkType);
 export const checkRequiredAndUrl = requiredAnd(checkUrl);
 export const checkRequiredAndArrayOf = requiredAnd(checkArrayOf);
+
+export { checkIsoDate } from './check-iso-date';