+++ packages/documentation/src/stories/misc/migration-guide/migrationv9-10.component.ts
@@ -5,7 +5,7 @@ import { V910Checks } from './types';
 import { _updateOnChange, _updatePersistedState } from './util/migration-checks.util';
 
 @customElement('migration-version-9-10')
-export class MigrationV99Component extends LitElement {
+export class MigrationV910Component extends LitElement {
   @property({ type: Number }) currentVersion?: number;
   @property({ type: String }) environment?: string;
   @property({ type: Boolean }) angular?: boolean;
@@ -21,7 +21,9 @@ export class MigrationV99Component extends LitElement {
     forms: {
       tooltip_validation: false,
       input_sizes: false,
+      select_empty: false,
       form_text: false,
+      card_control: false,
     },
     grid: {
       breakpoints: false,
@@ -37,6 +39,7 @@ export class MigrationV99Component extends LitElement {
       elevation: false,
       removed_spacing: false,
       renamed_spacing: false,
+      responsive_spacing: false,
       background: false,
       renamed_various_utilities: false,
       removed_various_utilities: false,
@@ -71,6 +74,7 @@ export class MigrationV99Component extends LitElement {
       stepper: false,
       dialog_icon: false,
       subnavigation: false,
+      product_card: false,
     },
     components: {
       alert: false,
@@ -79,6 +83,7 @@ export class MigrationV99Component extends LitElement {
       accordion_item_part: false,
       popover_trigger: false,
       tabs_anchor_navigation: false,
+      card_control: false,
     },
   };
 
@@ -162,11 +167,12 @@ export class MigrationV99Component extends LitElement {
             <p>
               Update Design System styles and components packages to version 10 by running these two
               commands in your project root:
-              <code languages="['bash']">npm install @swisspost/design-system-styles@10</code>
-              <code languages="['bash']">
-                npm install
-                @swisspost/design-system-components${this.angular ? '-angular' : nothing}@10
-              </code>
+              <code-block code=${'npm install @swisspost/design-system-styles@10'}></code-block>
+              <code-block
+                code=${this.angular
+                  ? 'npm install @swisspost/design-system-components-angular@10'
+                  : 'npm install @swisspost/design-system-components@10'}
+              ></code-block>
               ${!this.angular
                 ? html`
                     <p class="mt-8">
@@ -186,23 +192,28 @@ export class MigrationV99Component extends LitElement {
             <h3>Run Automigration Scripts 🪄</h3>
             <p>
               Many breaking changes can be fixed automatically using the
-              <code>@swisspost/design-system-eslint</code> package. Our very own, custom migration rules scan your
-              HTML and TypeScript files and apply fixes where possible. Each item marked with
-              <span class="tag tag-sm tag-info">🪄 migration rule</span> in the checklist below
-              is covered by one of them.
+              <code>@swisspost/design-system-eslint</code> package. Our very own, custom migration
+              rules scan your HTML and TypeScript files and apply fixes where possible. Each item
+              marked with <span class="tag tag-sm tag-info">🪄 migration rule</span> in the
+              checklist below is covered by one of them.
             </p>
             <ol>
               <li>
                 Install the Design System ESLint package as a dev dependency:
-                <code languages="['bash']">npm install @swisspost/design-system-eslint --save-dev</code>
+                <code-block
+                  code=${'npm install @swisspost/design-system-eslint --save-dev'}
+                ></code-block>
               </li>
               <li>
-                Run the migration rules using the official ESLint runner with the --fix flag at the root of your project:
-                <code languages="['bash']">npx eslint --config node_modules/@swisspost/design-system-eslint/dist/migrations.js --fix
-                </code>
+                Run the migration rules using the official ESLint runner with the --fix flag at the
+                root of your project:
+                <code-block
+                  code=${'npx eslint --config node_modules/@swisspost/design-system-eslint/dist/migrations.js --fix'}
+                ></code-block>
                 <span class="info">
-                  💡 This command applies migration rules using the official ESLint package without installing it as a project 
-                  dependency or modifying your existing ESLint configuration.
+                  💡 This command applies migration rules using the official ESLint package without
+                  installing it as a project dependency or modifying your existing ESLint
+                  configuration.
                 </span>
               </li>
               <li>
@@ -335,6 +346,7 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -349,6 +361,7 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -367,6 +380,7 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -388,6 +402,7 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -403,6 +418,7 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -434,6 +450,21 @@ export class MigrationV99Component extends LitElement {
                             </li>
                           </ul>
                         </span>
+                      </label>
+                    </div>
+                  </li>
+
+                  <li class="mb-16">
+                    <div class="form-check">
+                      <input
+                        id="components-card_control"
+                        class="form-check-input"
+                        type="checkbox"
+                        ?checked="${this.state.components.card_control}"
+                      />
+                      <label class="form-check-label" for="components-card_control">
+                        <code>post-card-control</code> component removed
+                        <span class="info"> Replace by the selection card. </span>
                       </label>
                     </div>
                   </li>
@@ -472,6 +503,7 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -498,6 +530,24 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
+                  <li class="mb-16">
+                    <div class="form-check">
+                      <input
+                        id="forms-select_empty"
+                        class="form-check-input"
+                        type="checkbox"
+                        ?checked="${this.state.forms.select_empty}"
+                      />
+                      <label class="form-check-label" for="forms-select_empty">
+                        <code>.form-select-empty</code> class deprecated
+                        <span class="info">
+                          This class has been replaced by a modern CSS selector, which detects the
+                          presence of an empty option.
+                        </span>
+                      </label>
+                    </div>
+                  </li>
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -512,6 +562,25 @@ export class MigrationV99Component extends LitElement {
                         >
                         <code>.form-text</code> class has been renamed to
                         <code>.form-hint</code>
+                      </label>
+                    </div>
+                  </li>
+
+                  <li class="mb-16">
+                    <div class="form-check">
+                      <input
+                        id="styles-card_control"
+                        class="form-check-input"
+                        type="checkbox"
+                        ?checked="${this.state.forms.card_control}"
+                      />
+                      <label class="form-check-label" for="styles_card_control">
+                        CSS classes for Standard HTML <code>card-control</code> component renamed
+                        <span class="info">
+                          Replace <code>.checkbox-button-card</code> and
+                          <code>.radio-button-card</code> with
+                          <code>.selection-card</code> component.
+                        </span>
                       </label>
                     </div>
                   </li>
@@ -805,6 +874,46 @@ export class MigrationV99Component extends LitElement {
                           <li><code>*-small-giant</code> is now <code>*-78</code></li>
                           <li><code>*-giant</code> is now <code>*-80</code></li>
                         </ul>
+                      </label>
+                    </div>
+                  </li>
+                  <li class="mb-16">
+                    <div class="form-check">
+                      <input
+                        id="utilities-responsive_spacing"
+                        class="form-check-input"
+                        type="checkbox"
+                        ?checked="${this.state.utilities.responsive_spacing}"
+                      />
+                      <label class="form-check-label" for="utilities-responsive_spacing">
+                        <span data-info="automigration" class="tag tag-sm tag-info"
+                          >🪄 migration rule</span
+                        >
+                        Responsive margin, padding and gap utility classes renamed
+                        <ul>
+                          <li><code>*-tiny-r</code> is now <code>*-12 *-md-16</code></li>
+                          <li><code>*-small-r</code> is now <code>*-12 *-sm-16</code></li>
+                          <li><code>*-regular-r</code> is now <code>*-16 *-md-24</code></li>
+                          <li>
+                            <strong><code>*-large-r</code></strong> is now
+                            <strong><code>*-16 *-md-24 *-lg-32</code></strong>
+                          </li>
+                          <li>
+                            <strong><code>*-big-r</code></strong> is now
+                            <strong><code>*-24 *-md-32 *-lg-40</code></strong>
+                          </li>
+                          <li><code>*-bigger-big-r</code> is now <code>*-24 *-md-48</code></li>
+                          <li>
+                            <strong><code>*-huge-r</code></strong> is now
+                            <strong><code>*-32 *-md-40 *-lg-56</code></strong>
+                          </li>
+                          <li><code>*-giant-r</code> is now <code>*-56 *-md-80</code></li>
+                        </ul>
+                        <span class="info"
+                          >⚠️ Bold entries (<code>large</code>, <code>big</code>, <code>huge</code>)
+                          are not 1:1 migrations, you should carefully review the autofix output to
+                          ensure spacing behaves as intended at each breakpoint.</span
+                        >
                       </label>
                     </div>
                   </li>
@@ -1556,6 +1665,7 @@ export class MigrationV99Component extends LitElement {
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -1565,13 +1675,14 @@ export class MigrationV99Component extends LitElement {
                         ?checked="${this.state.others.stepper}"
                       />
                       <label class="form-check-label" for="others.stepper">
-                        Removed the stepper HTML component.
+                        Stepper HTML component removed
                         <span class="info"
                           >You can now use the <code>post-stepper</code> web component.</span
                         >
                       </label>
                     </div>
                   </li>
+
                   <li class="mb-16">
                     <div class="form-check">
                       <input
@@ -1602,6 +1713,24 @@ export class MigrationV99Component extends LitElement {
                         Subnavigation component removed
                         <span class="info">
                           Replaced by the <code>post-tabs</code> component.
+                        </span>
+                      </label>
+                    </div>
+                  </li>
+
+                  <li class="mb-16">
+                    <div class="form-check">
+                      <input
+                        id="others-product_card"
+                        class="form-check-input"
+                        type="checkbox"
+                        ?checked="${this.state.others.product_card}"
+                      />
+                      <label class="form-check-label" for="others-product_card">
+                        Card product component removed.
+                        <span class="info">
+                          The <code>.product-card</code> and
+                          <code>.product-navigation</code> classes are no longer effective.
                         </span>
                       </label>
                     </div>
@@ -1637,7 +1766,11 @@ export class MigrationV99Component extends LitElement {
               <li>
                 Once you've verified that your project builds and displays correctly, uninstall the
                 packages by running:
-                <code languages="['bash']">npm uninstall bootstrap @ng-bootstrap/ng-bootstrap</code>
+                <code-block
+                  code=${this.angular
+                    ? 'npm uninstall bootstrap @ng-bootstrap/ng-bootstrap'
+                    : 'npm uninstall bootstrap'}
+                ></code-block>
               </li>
             </ol>
           </li>