+++ packages/components-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @swisspost/design-system-components-react
 
+## 10.0.0-next.66
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-components@10.0.0-next.66
+
+## 10.0.0-next.65
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-components@10.0.0-next.65
+
 ## 10.0.0-next.64
 
 ### Patch Changes